Get font and workmark working
Helvetica is the official font of the government of Canada,
but it's a proprietary font, requiring a licence for use.
This has led to developers to create pictures of the font,
using pngs or svgs which create a bunch of usability and
 accessibility problems.
Here we're taking a different approach: Using a
helvetica compatible font instead.