Use full path to avoid translation bugs in url
Previously navigating to `/drug-seizures` would render a link to `new`.
If you click it, you would navigate to `/drug-seizures/new`.
If you switched language to French it would rerender the link pointing to `nouveau`.
If you click it, you would navigate to `/drug-seizures/nouveau`, a combination that does not exist.